Clickable Card URLs and Menu Toggling
To enhance user-friendliness, the card URLs on the pages have been updated to be clickable. When users click on a card, it will now open the corresponding payment link in a new tab or window, making it convenient for them to access payment options for various services. -An issue that previously hindered the smooth functioning of the page's menu on youtube.html has been successfully resolved. Users can now toggle the menu with ease, allowing for seamless navigation and access to different sections of the page